% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials_parameters.R
\name{covariance_function}
\alias{covariance_function}
\title{Covariance function type (qualitative tuning parameter)}
\usage{
covariance_function(
  values = c("exponential", "spherical", "gaussian", "matern")
)
}
\arguments{
\item{values}{Character vector of allowable covariance function names.
Defaults to the four most common choices.}
}
\value{
A \code{dials} qualitative parameter object.
}
\description{
Covariance function type (qualitative tuning parameter)
}
\examples{
covariance_function()
}
