% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_geometry}
\alias{drop_geometry}
\title{Drop geometry and return a plain data.frame}
\usage{
drop_geometry(data)
}
\arguments{
\item{data}{An \code{sf} object or \code{data.frame}.}
}
\value{
A \code{data.frame} (or \code{tibble}) without a geometry column.
}
\description{
Drop geometry and return a plain data.frame
}
\keyword{internal}
