% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bundle.R
\name{get_result_types}
\alias{get_result_types}
\title{Get allowed result keys from the database}
\usage{
get_result_types(db_path = NULL, read_only = TRUE)
}
\arguments{
\item{db_path}{Optional path to the database.}

\item{read_only}{Logical; open connection in read-only mode.}
}
\value{
Character vector of result keys.
}
\description{
Returns unique result keys defined in \code{key_spec}.
}
\keyword{internal}
