% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopher-package.R
\docType{package}
\name{gopher-package}
\alias{gopher}
\alias{gopher-package}
\title{gopher: tidymodels-Compatible Gaussian Process Model Zoo}
\description{
A gateway to various Gaussian Process (GP) packages available on CRAN, including gstat, fields, GPvecchia, spNNGP, and PrestoGP. Provides a unified tidymodels-native interface through parsnip for model parameter adjustment and hyperparameter tuning. Supports spatial-only and spatiotemporal Kriging, sf inputs, prediction capabilities, and covariates for residual Gaussian Process (Universal Kriging) models. Parameters are automatically matched between the unified Kriging-centered argument names in gstat and the formulaic arguments used in other packages for the same construct.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sigmafelix/gopher}
  \item Report bugs at \url{https://github.com/sigmafelix/gopher/issues}
}

}
\author{
\strong{Maintainer}: Insang Song \email{geoissong@gmail.com}

Authors:
\itemize{
  \item C Codex
}

}
\keyword{internal}
