% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gopher_con}
\alias{gopher_con}
\title{Open a database connection}
\usage{
gopher_con(db_path = NULL, read_only = FALSE)
}
\arguments{
\item{db_path}{Full path to the database file. If \code{NULL}, resolved via
\code{\link[=gopher_db_path]{gopher_db_path()}}.}

\item{read_only}{Logical. If \code{TRUE}, opens in read-only mode.}
}
\value{
A DBI connection.
}
\description{
Open a database connection
}
