% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gopher_query}
\alias{gopher_query}
\title{Run a parameterized query and return a data frame}
\usage{
gopher_query(con, sql, params = NULL)
}
\arguments{
\item{con}{A DBI connection.}

\item{sql}{A SQL query string, using \verb{?} placeholders for parameters.}

\item{params}{A list of parameters to bind.}
}
\value{
A data.frame.
}
\description{
Run a parameterized query and return a data frame
}
