% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials_parameters.R
\name{gp_nugget}
\alias{gp_nugget}
\title{GP nugget variance (quantitative tuning parameter)}
\usage{
gp_nugget(range = c(0, 5), trans = NULL)
}
\arguments{
\item{range}{A numeric vector of length 2 giving lower and upper bounds.}

\item{trans}{A \code{scales} transformation, or \code{NULL}.}
}
\value{
A \code{dials} quantitative parameter object.
}
\description{
GP nugget variance (quantitative tuning parameter)
}
\examples{
gp_nugget()
}
