% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials_parameters.R
\name{gp_range}
\alias{gp_range}
\title{GP range parameter (quantitative tuning parameter)}
\usage{
gp_range(range = c(0.001, 1000), trans = NULL)
}
\arguments{
\item{range}{A numeric vector of length 2 giving the lower and upper bounds
of the range parameter search space. Values are on the original scale.}

\item{trans}{A \code{scales} transformation object, or \code{NULL} (default).}
}
\value{
A \code{dials} quantitative parameter object.
}
\description{
GP range parameter (quantitative tuning parameter)
}
\examples{
gp_range()
}
