% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials_parameters.R
\name{gp_sill}
\alias{gp_sill}
\title{GP partial sill variance (quantitative tuning parameter)}
\usage{
gp_sill(range = c(1e-06, 1000), trans = NULL)
}
\arguments{
\item{range}{A numeric vector of length 2 giving lower and upper bounds.}

\item{trans}{A \code{scales} transformation, or \code{NULL} (default).}
}
\value{
A \code{dials} quantitative parameter object.
}
\description{
GP partial sill variance (quantitative tuning parameter)
}
\examples{
gp_sill()
}
