% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bundle.R
\name{insert_object_files_with_con}
\alias{insert_object_files_with_con}
\title{Insert object_files into database (with connection)}
\usage{
insert_object_files_with_con(file_data, con)
}
\arguments{
\item{file_data}{Data frame of object_files to insert}

\item{con}{A DBI connection to the database}
}
\value{
Number of rows inserted
}
\description{
Inserts validated object_file records into the object_file table.
}
\keyword{internal}
