% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bundle.R
\name{insert_objects_with_con}
\alias{insert_objects_with_con}
\title{Insert objects into the database (with connection)}
\usage{
insert_objects_with_con(object_data, con)
}
\arguments{
\item{object_data}{Data frame of validated objects.}

\item{con}{A DBI connection to the database.}
}
\value{
Invisibly returns the number of rows inserted.
}
\description{
Inserts validated object data into the object table using an existing connection.
}
\keyword{internal}
