% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bundle.R
\name{insert_people_with_con}
\alias{insert_people_with_con}
\title{Insert people into database (with connection)}
\usage{
insert_people_with_con(people_data, con)
}
\arguments{
\item{people_data}{Data frame of people to insert}

\item{con}{A DBI connection to the database}
}
\value{
List with n_inserted count
}
\description{
Inserts validated people records into the people table.
}
\keyword{internal}
