% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bundle.R
\name{insert_results_with_con}
\alias{insert_results_with_con}
\title{Insert results into database (with connection)}
\usage{
insert_results_with_con(result_data, con)
}
\arguments{
\item{result_data}{Data frame of results to insert}

\item{con}{A DBI connection to the database}
}
\value{
Number of rows inserted
}
\description{
Inserts validated result records into the result table.
}
\keyword{internal}
