% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoGP_engine.R
\name{PrestoGP_gp_predict}
\alias{PrestoGP_gp_predict}
\title{Predict from a PrestoGP-fitted Gaussian Process model}
\usage{
PrestoGP_gp_predict(
  object,
  new_data,
  type = "numeric",
  level = 0.95,
  m_pred = NULL,
  coord_cols = NULL,
  time_col = NULL,
  time_scale = 1,
  ...
)
}
\arguments{
\item{object}{A \code{"gopher_PrestoGP_fit"} returned by \code{\link[=PrestoGP_gp_fit]{PrestoGP_gp_fit()}}.}

\item{new_data}{An \code{sf} object or \code{data.frame} with coordinates.}

\item{type}{\code{"numeric"} (default) or \code{"pred_int"}.}

\item{level}{Confidence level for prediction intervals (default \code{0.95}).}

\item{m_pred}{Optional number of neighbors for prediction.}

\item{coord_cols}{Character(2) coordinate column names for non-sf input.}

\item{time_col}{Optional character scalar specifying a time column for
spatiotemporal modelling.}

\item{time_scale}{Numeric scalar used to rescale time when \code{time_col} is
provided. Default \code{1}.}

\item{...}{Forwarded to \code{PrestoGP::prestogp_predict()}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with prediction columns.
}
\description{
Predict from a PrestoGP-fitted Gaussian Process model
}
