% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{uid_native}
\alias{uid_native}
\title{Extract native id from a UID like "mag:M001" (R-side)}
\usage{
uid_native(uid)
}
\arguments{
\item{uid}{}
}
\description{
Extract native id from a UID like "mag:M001" (R-side)
}
