% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bundle.R
\name{validate_people_with_con}
\alias{validate_people_with_con}
\title{Validate people against database (with connection)}
\usage{
validate_people_with_con(people_data, con)
}
\arguments{
\item{people_data}{Data frame of people to validate}

\item{con}{A DBI connection to the database}
}
\value{
List with valid (logical) and message (character) elements
}
\description{
Checks that person_ids are unique within bundle and don't already exist
in database.
}
\keyword{internal}
