% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{with_gopher_con}
\alias{with_gopher_con}
\title{Evaluate a function with a database connection}
\usage{
with_gopher_con(.f, ..., db_path = NULL, read_only = FALSE)
}
\arguments{
\item{.f}{A function with first argument \code{con}.}

\item{...}{Additional arguments passed to \code{.f}.}

\item{db_path, read_only}{Passed to \code{\link[=gopher_con]{gopher_con()}}.}
}
\value{
The return value of \code{.f}.
}
\description{
Evaluate a function with a database connection
}
