% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gopher_con}
\alias{gopher_con}
\title{Open a database connection}
\usage{
gopher_con(db_path = NULL, read_only = FALSE)
}
\arguments{
\item{db_path}{path to gopher DB}

\item{read_only}{Logical. If \code{TRUE}, attempts to open the SQLite DB in read-only mode.}
}
\value{
A DBI connection.
}
\description{
Open a database connection
}
