% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gopher_db_path}
\alias{gopher_db_path}
\title{Get path to gopheR database}
\usage{
gopher_db_path(path = here::here("data"), db = "gopheR_db.sqlite")
}
\arguments{
\item{path}{(\emph{string}) path to db}

\item{db}{(\emph{string}) database name}
}
\value{
file path to gopheR_db.sqlite
}
\description{
Get path to gopheR database
}
