% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{gopher_db_path}
\alias{gopher_db_path}
\title{Get path to a gopheR SQLite database file}
\usage{
gopher_db_path(
  path = NULL,
  db = getOption("gopheR.db_file", "gopheR_db.sqlite")
)
}
\arguments{
\item{path}{Character scalar. Directory containing the database file.
If \code{NULL}, uses \code{getOption("gopheR.db_path")} or \code{Sys.getenv("GOPHER_DB_PATH")}.}

\item{db}{Character scalar. Database filename.}
}
\value{
A length-1 character vector: full path to the database file.
}
\description{
Reliable set of code to determine where exactly a gopheR database is in a
project.
}
